src/requirements.txt
requests
beautifulsoup4
tqdm
tabulate
pyyaml
python-dotenv
